Adjust to linux-support-4.9.0-4
This package uses some code from the linux-support package to generate
files. The way in which this done in the packaging is similar to how the
linux source package is managed.

debian/rules.defs defines a KERNELVERSION which requires you to have the
linux-support-$KERNELVERSION package installed.

For now, update it to 4.9.0-4, since that's the current latest in
Stretch. I'm working on the packaging on a stretch system.

After doing so...
    -$ make -f debian/rules debian/control-real
    [...]
      File "/usr/src/linux-support-4.9.0-4/lib/python/debian_linux/gencontrol.py", line 124, in do_main_recurse
          for featureset in self.config['base', ]['featuresets']:
    KeyError: 'featuresets'

Apparently the newer linux-support requires featuresets to be set. We
simply add it with only the option none, since we don't use it.