d/t/control-*: bump grub config incompat for now
In 8c378bc011 I moved the grub config file that was wrongly placed
inside the hypervisor package, and assumed that it would be fixed in a
stable update, and that the orig tar would not change before that.
However, we don't know yet what will happen in stretch security.

So, don't assume the fix would happen in a specific version in stable.
For now, bump the cutoff point all the way up to 4.10 so we can lower it
again when the actual fix did happen in stretch.

The end goal is to be able to upgrade from stretch to buster without
this problem.  As long as the fix didn't happen in stable, and we have a
newer xen in unstable, the upgrade will hit the conflict.

Signed-off-by: Hans van Kranenburg <hans@knorrie.org>